+++ pytest_aiohttp/__init__.py
@@ -1,4 +1,1 @@
-__version__ = '0.3.0'
-
-
-from aiohttp.pytest_plugin import *
+from ._version import version as __version__  # noqa